[explorer][delegation] add status icon to my deposits section
