+++ src/pages/DelegatoryValidator/MyDepositsSection.tsx
@@ -1,5 +1,6 @@
 import {
   Button,
+  Chip,
   Stack,
   Table,
   TableHead,
@@ -18,6 +19,92 @@ import GeneralTableHeaderCell from "../../components/Table/GeneralTableHeaderCel
 import GeneralTableRow from "../../components/Table/GeneralTableRow";
 import MyDepositsStatusTooltip from "./Components/MyDepositsStatusTooltip";
 import {getLockedUtilSecs} from "./utils";
+import LockIcon from "@mui/icons-material/Lock";
+import PendingIcon from "@mui/icons-material/Pending";
+import CheckCircleIcon from "@mui/icons-material/CheckCircle";
+import {
+  STAKED_BACKGROUND_COLOR_DARK,
+  STAKED_BACKGROUND_COLOR_LIGHT,
+  STAKED_DESCRIPTION,
+  STAKED_LABEL,
+  STAKED_TEXT_COLOR_DARK,
+  STAKED_TEXT_COLOR_LIGHT,
+  WITHDRAW_PENDING_BACKGROUND_COLOR_DARK,
+  WITHDRAW_PENDING_BACKGROUND_COLOR_LIGHT,
+  WITHDRAW_PENDING_DESCRIPTION,
+  WITHDRAW_PENDING_LABEL,
+  WITHDRAW_PENDING_TEXT_COLOR_DARK,
+  WITHDRAW_PENDING_TEXT_COLOR_LIGHT,
+  WITHDRAW_READY_BACKGROUND_COLOR_DARK,
+  WITHDRAW_READY_BACKGROUND_COLOR_LIGHT,
+  WITHDRAW_READY_DESCRIPTION,
+  WITHDRAW_READY_LABEL,
+  WITHDRAW_READY_TEXT_COLOR_DARK,
+  WITHDRAW_READY_TEXT_COLOR_LIGHT,
+} from "./constants";
+
+enum StakingStatusStep {
+  "staked",
+  "withdrawPending",
+  "withdrawReady",
+}
+type Steps = keyof typeof StakingStatusStep;
+
+export interface StakingStatusStepInterface {
+  label: string;
+  description: string;
+  icon: JSX.Element;
+  sxLight: {
+    color: string;
+    backgroundColor: string;
+  };
+  sxDark: {
+    color: string;
+    backgroundColor: string;
+  };
+}
+
+const steps = [
+  {
+    label: STAKED_LABEL,
+    description: STAKED_DESCRIPTION,
+    icon: <LockIcon />,
+    sxLight: {
+      color: STAKED_TEXT_COLOR_LIGHT,
+      backgroundColor: STAKED_BACKGROUND_COLOR_LIGHT,
+    },
+    sxDark: {
+      color: STAKED_TEXT_COLOR_DARK,
+      backgroundColor: STAKED_BACKGROUND_COLOR_DARK,
+    },
+  },
+  {
+    label: WITHDRAW_PENDING_LABEL,
+    description: WITHDRAW_PENDING_DESCRIPTION,
+    icon: <PendingIcon />,
+    sxLight: {
+      color: WITHDRAW_PENDING_TEXT_COLOR_LIGHT,
+      backgroundColor: WITHDRAW_PENDING_BACKGROUND_COLOR_LIGHT,
+    },
+    sxDark: {
+      color: WITHDRAW_PENDING_TEXT_COLOR_DARK,
+      backgroundColor: WITHDRAW_PENDING_BACKGROUND_COLOR_DARK,
+    },
+  },
+  {
+    label: WITHDRAW_READY_LABEL,
+    description: WITHDRAW_READY_DESCRIPTION,
+    icon: <CheckCircleIcon />,
+    sxLight: {
+      color: WITHDRAW_READY_TEXT_COLOR_LIGHT,
+      backgroundColor: WITHDRAW_READY_BACKGROUND_COLOR_LIGHT,
+    },
+    sxDark: {
+      color: WITHDRAW_READY_TEXT_COLOR_DARK,
+      backgroundColor: WITHDRAW_READY_BACKGROUND_COLOR_DARK,
+    },
+  },
+];
 
 type MyDepositsSectionProps = {
   accountResource?: Types.MoveResource | undefined;
@@ -65,7 +152,20 @@ function AmountCell({}: MyDepositsSectionProps) {
 }
 
 function StatusCell({}: MyDepositsSectionProps) {
-  return <GeneralTableCell>Staked</GeneralTableCell>;
+  const theme = useTheme();
+
+  // TODO(jill): temp workaround since we don't have status data yet
+  const step = steps[0];
+  return (
+    <GeneralTableCell>
+      <Chip
+        icon={step.icon}
+        label={step.label}
+        sx={theme.palette.mode === "dark" ? step.sxDark : step.sxLight}
+        color="primary"
+      />
+    </GeneralTableCell>
+  );
 }
 
 function UnlockDateCell({accountResource}: MyDepositsSectionProps) {
@@ -118,7 +218,13 @@ export default function MyDepositsSection({
                 }}
                 header={MyDepositsHeader[columnName]}
                 key={idx}
-                tooltip={columnName === "status" && <MyDepositsStatusTooltip />}
+                tooltip={
+                  columnName === "status" && (
+                    <MyDepositsStatusTooltip
+                      steps={steps as [StakingStatusStepInterface]}
+                    />
+                  )
+                }
               />
             ))}
           </TableRow>
